Add files via upload
the directional movement wasn't working properly. It would move around the array but it was ignoring the exits since they weren't connected to the scenes themselves. ex. you could move east even if east wasn't 'open' as long as there was a scene in the adjacent array index. and if there wasn't it would crash. Now the direction objects, of class Dir are set in the Scene constructor, and on input it checks for directions attached to currentScene.